--- 2019-02-10/cutils.h
@@ -1,226 +1,0 @@
-/*
- * C utilities
- * 
- * Copyright (c) 2016 Fabrice Bellard
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-#ifndef CUTILS_H
-#define CUTILS_H
-
-#include <inttypes.h>
-
-#define likely(x)       __builtin_expect(!!(x), 1)
-#define unlikely(x)     __builtin_expect(!!(x), 0)
-#define force_inline inline __attribute__((always_inline))
-#define no_inline __attribute__((noinline))
-#define __maybe_unused __attribute__((unused))
-
-#define xglue(x, y) x ## y
-#define glue(x, y) xglue(x, y)
-#define stringify(s)    tostring(s)
-#define tostring(s)     #s
-
-#ifndef offsetof
-#define offsetof(type, field) ((size_t) &((type *)0)->field)
-#endif
-#define countof(x) (sizeof(x) / sizeof(x[0]))
-
-#define DLL_PUBLIC __attribute__ ((visibility ("default")))
-
-#ifndef _BOOL_defined
-#define _BOOL_defined
-#undef FALSE
-#undef TRUE
-
-typedef int BOOL;
-enum {
-    FALSE = 0,
-    TRUE = 1,
-};
-#endif
-
-/* this test works at least with gcc */
-#if defined(__SIZEOF_INT128__)
-#define HAVE_INT128
-#endif
-
-#ifdef HAVE_INT128
-typedef __int128 int128_t;
-typedef unsigned __int128 uint128_t;
-#endif
-
-static inline int max_int(int a, int b)
-{
-    if (a > b)
-        return a;
-    else
-        return b;
-}
-
-static inline int min_int(int a, int b)
-{
-    if (a < b)
-        return a;
-    else
-        return b;
-}
-
-void *mallocz(size_t size);
-
-#if defined(_WIN32)
-static inline uint16_t bswap_16(uint16_t v)
-{
-    return (((v) & 0xff) << 8) | ((v) >> 8);
-}
-static inline uint32_t bswap_32(uint32_t v)
-{
-    return ((v & 0xff000000) >> 24) | ((v & 0x00ff0000) >>  8) |
-        ((v & 0x0000ff00) <<  8) | ((v & 0x000000ff) << 24);
-}
-#define bswap_64(x) _byteswap_uint64(x)
-#elif defined(__APPLE__)
-#include <libkern/OSByteOrder.h>
-#define bswap_16(x) OSSwapInt16(x)
-#define bswap_32(x) OSSwapInt32(x)
-#define bswap_64(x) OSSwapInt64(x)
-#else
-#include <byteswap.h>
-#endif
-
-static inline uint16_t get_le16(const uint8_t *ptr)
-{
-    return ptr[0] | (ptr[1] << 8);
-}
-
-static inline uint32_t get_le32(const uint8_t *ptr)
-{
-    return ptr[0] | (ptr[1] << 8) | (ptr[2] << 16) | (ptr[3] << 24);
-}
-
-static inline uint64_t get_le64(const uint8_t *ptr)
-{
-    return get_le32(ptr) | ((uint64_t)get_le32(ptr + 4) << 32);
-}
-
-static inline void put_le16(uint8_t *ptr, uint16_t v)
-{
-    ptr[0] = v;
-    ptr[1] = v >> 8;
-}
-
-static inline void put_le32(uint8_t *ptr, uint32_t v)
-{
-    ptr[0] = v;
-    ptr[1] = v >> 8;
-    ptr[2] = v >> 16;
-    ptr[3] = v >> 24;
-}
-
-static inline void put_le64(uint8_t *ptr, uint64_t v)
-{
-    put_le32(ptr, v);
-    put_le32(ptr + 4, v >> 32);
-}
-
-static inline uint32_t get_be32(const uint8_t *d)
-{
-    return (d[0] << 24) | (d[1] << 16) | (d[2] << 8) | d[3];
-}
-
-static inline void put_be32(uint8_t *d, uint32_t v)
-{
-    d[0] = v >> 24;
-    d[1] = v >> 16;
-    d[2] = v >> 8;
-    d[3] = v >> 0;
-}
-
-static inline void put_be64(uint8_t *d, uint64_t v)
-{
-    put_be32(d, v >> 32);
-    put_be32(d + 4, v);
-}
-
-#ifdef WORDS_BIGENDIAN
-static inline uint32_t cpu_to_be32(uint32_t v)
-{
-    return v;
-}
-#else
-static inline uint32_t cpu_to_be32(uint32_t v)
-{
-    return bswap_32(v);
-}
-#endif
-
-static inline int ctz32(uint32_t val)
-{
-#if defined(__GNUC__) && __GNUC__ >= 4
-    return val ? __builtin_ctz(val) : 32;
-#else
-    /* Binary search for the trailing one bit.  */
-    int cnt;
-    cnt = 0;
-    if (!(val & 0x0000FFFFUL)) {
-        cnt += 16;
-        val >>= 16;
-    }
-    if (!(val & 0x000000FFUL)) {
-        cnt += 8;
-        val >>= 8;
-    }
-    if (!(val & 0x0000000FUL)) {
-        cnt += 4;
-        val >>= 4;
-    }
-    if (!(val & 0x00000003UL)) {
-        cnt += 2;
-        val >>= 2;
-    }
-    if (!(val & 0x00000001UL)) {
-        cnt++;
-        val >>= 1;
-    }
-    if (!(val & 0x00000001UL)) {
-        cnt++;
-    }
-    return cnt;
-#endif
-}
-
-
-void *mallocz(size_t size);
-void pstrcpy(char *buf, int buf_size, const char *str);
-char *pstrcat(char *buf, int buf_size, const char *s);
-int strstart(const char *str, const char *val, const char **ptr);
-
-typedef struct {
-    uint8_t *buf;
-    size_t size;
-    size_t allocated_size;
-} DynBuf;
-
-void dbuf_init(DynBuf *s);
-void dbuf_write(DynBuf *s, size_t offset, const uint8_t *data, size_t len);
-void dbuf_putc(DynBuf *s, uint8_t c);
-void dbuf_putstr(DynBuf *s, const char *str);
-void dbuf_free(DynBuf *s);
-
-#endif /* CUTILS_H */
