+++ cutils.h
@@ -85,11 +85,21 @@ static inline int min_int(int a, int b)
 void *mallocz(size_t size);
 
 #if defined(_WIN32)
+static inline uint16_t bswap_16(uint16_t v)
+{
+    return (((v) & 0xff) << 8) | ((v) >> 8);
+}
 static inline uint32_t bswap_32(uint32_t v)
 {
     return ((v & 0xff000000) >> 24) | ((v & 0x00ff0000) >>  8) |
         ((v & 0x0000ff00) <<  8) | ((v & 0x000000ff) << 24);
 }
+#define bswap_64(x) _byteswap_uint64(x)
+#elif defined(__APPLE__)
+#include <libkern/OSByteOrder.h>
+#define bswap_16(x) OSSwapInt16(x)
+#define bswap_32(x) OSSwapInt32(x)
+#define bswap_64(x) OSSwapInt64(x)
 #else
 #include <byteswap.h>
 #endif
@@ -160,17 +170,39 @@ static inline uint32_t cpu_to_be32(uint32_t v)
 }
 #endif
 
-/* XXX: optimize */
-static inline int ctz32(uint32_t a)
+static inline int ctz32(uint32_t val)
 {
-    int i;
-    if (a == 0)
-        return 32;
-    for(i = 0; i < 32; i++) {
-        if ((a >> i) & 1)
-            return i;
+#if defined(__GNUC__) && __GNUC__ >= 4
+    return val ? __builtin_ctz(val) : 32;
+#else
+    /* Binary search for the trailing one bit.  */
+    int cnt;
+    cnt = 0;
+    if (!(val & 0x0000FFFFUL)) {
+        cnt += 16;
+        val >>= 16;
     }
-    return 32;
+    if (!(val & 0x000000FFUL)) {
+        cnt += 8;
+        val >>= 8;
+    }
+    if (!(val & 0x0000000FUL)) {
+        cnt += 4;
+        val >>= 4;
+    }
+    if (!(val & 0x00000003UL)) {
+        cnt += 2;
+        val >>= 2;
+    }
+    if (!(val & 0x00000001UL)) {
+        cnt++;
+        val >>= 1;
+    }
+    if (!(val & 0x00000001UL)) {
+        cnt++;
+    }
+    return cnt;
+#endif
 }
